:internal: delete and backspace on item layer is now controlled by the scene. This has advantages deleting the items

+++ egCAS/src/view/egcasscene.cpp
@@ -457,10 +457,25 @@ void EgCasScene::keyPressEvent(QKeyEvent* keyEvent)
                 QList<QGraphicsItem*> list = selectedItems();
                 QGraphicsItem *item;
                 foreach (item, list) {
-                        deleteItem(item);
-                        m_document.itemDeleted(item);
+                        if (    item->type() == static_cast<int>(EgcGraphicsItemType::EgcFormulaItemType)
+                             || item->type() == static_cast<int>(EgcGraphicsItemType::EgcPixmapItemType)
+                             || item->type() == static_cast<int>(EgcGraphicsItemType::EgcTextItemType)) {
+                                EgcFormulaItem *formula = dynamic_cast<EgcFormulaItem*>(item);
+                                if (formula) {  //formula
+                                        if (formula->aboutToBeDeleted()) {
+                                                deleteItem(item);
+                                                m_document.itemDeleted(item);
+                                                keyEvent->accept();
+                                        } else {  //call standard version that triggers the elements key event routine
+                                                QGraphicsScene::keyPressEvent(keyEvent);
+                                        }
+                                } else {  // pixmap or text
+                                        deleteItem(item);
+                                        m_document.itemDeleted(item);
+                                        keyEvent->accept();
+                                }
+                        }
                 }
-                keyEvent->accept();
         } else {
                 QGraphicsScene::keyPressEvent(keyEvent);
         }
